sentry bugs

+++ app/Http/Controllers/AuditController.php
@@ -1029,12 +1029,14 @@ class AuditController extends Controller
         $amenity_inspection = AmenityInspection::where('id', '=', $amenity_id)->first();
         $ordering_buildings = OrderingBuilding::where('audit_id', '=', $audit_id)->where('amenity_inspection_id', '=', $amenity_id)->first();
         $cached_building    = CachedBuilding::where('audit_id', '=', $audit_id)->where('amenity_inspection_id', '=', $amenity_id)->first();
-        $project_amenity    = ProjectAmenity::where('project_id', '=', $project_id)->where('amenity_id', '=', $amenity_inspection->amenity_id)->first();
+        if($amenity_inspection) {
+        	$project_amenity    = ProjectAmenity::where('project_id', '=', $project_id)->where('amenity_id', '=', $amenity_inspection->amenity_id)->first();
+        	 $amenity_inspection->delete();
+	         $project_amenity->delete();
+        }
 
-        $amenity_inspection->delete();
         $ordering_buildings->delete();
         $cached_building->delete();
-        $project_amenity->delete();
 
         $new_comment = new Comment([
           'user_id'       => Auth::user()->id,
